fix: api 통신 로직 변경해 불필요하게 + 의도와는 다르게 RDD 테이블의 is_success가 바뀌는 오류 해결
